Save work on Windows port (Issue #116)

+++ pappl/printer-raw.c
@@ -119,7 +119,7 @@ _papplPrinterRunRaw(
 
           // Accept the connection...
           sockaddrlen = sizeof(sockaddr);
-          if ((sock = accept(printer->raw_listeners[i].fd, (struct sockaddr *)&sockaddr, &sockaddrlen)) < 0)
+          if ((sock = (int)accept(printer->raw_listeners[i].fd, (struct sockaddr *)&sockaddr, &sockaddrlen)) < 0)
           {
             papplLogPrinter(printer, PAPPL_LOGLEVEL_ERROR, "Unable to accept socket print connection: %s", strerror(errno));
             continue;
